// src/app/Types/ContactTypes.ts
export interface UserMessageProps {
  [key: string]: string;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  message: string;
}


export interface ContactFormErrorProps {
  [key: string]: string;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  message: string;
}